Change package name to Advanced Android course number
