+++ app/src/androidTest/java/csc494/finalProject/SilentSpots/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package csc415.finalProject.SilentSpots;
+package csc494.finalProject.SilentSpots;
 
 import android.content.Context;
