+++ app/src/main/java/csc494/finalProject/SilentSpots/RecyclerItemClickListener.java
@@ -1,4 +1,4 @@
-package csc415.finalProject.SilentSpots;
+package csc494.finalProject.SilentSpots;
 
 import android.content.Context;
 import android.view.GestureDetector;
