+++ app/src/main/java/csc494/finalProject/SilentSpots/Rule.java
@@ -1,4 +1,4 @@
-package csc415.finalProject.SilentSpots;
+package csc494.finalProject.SilentSpots;
 
 import com.google.firebase.firestore.GeoPoint;
