+++ app/src/test/java/csc494/finalProject/SilentSpots/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package csc415.finalProject.SilentSpots;
+package csc494.finalProject.SilentSpots;
 
 import org.junit.Test;
